requirements.txt
# Core ML dependencies
tensorflow>=2.13.0,<3.0.0
numpy>=2.0.0

# Data processing and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Web frameworks
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
gradio>=4.0.0

# Utilities
python-multipart>=0.0.6
pydantic>=2.0.0

# Optional: Development tools
# jupyter>=1.0.0
# ipykernel>=6.25.0
# pytest>=7.4.0
# black>=23.7.0
# flake8>=6.1.0
